added a base class for Mutation table

added results view mutation table component (not fully implemented yet)


added allele frequency columns


renamed MSKTable references to LazyMobXTable
