+++ src/pages/patientView/copyNumberAlterations/CopyNumberTableWrapper.tsx
@@ -6,17 +6,16 @@ import LazyMobXTable from "shared/components/lazyMobXTable/LazyMobXTable";
 import {DiscreteCopyNumberData} from "shared/api/generated/CBioPortalAPI";
 import {Column} from "shared/components/lazyMobXTable/LazyMobXTable";
 import * as _ from 'lodash';
-import MrnaExprColumnFormatter from "../mutation/column/MrnaExprColumnFormatter";
+import MrnaExprColumnFormatter from "shared/components/mutationTable/column/MrnaExprColumnFormatter";
 import CohortColumnFormatter from "./column/CohortColumnFormatter";
 import CnaColumnFormatter from "./column/CnaColumnFormatter";
 import AnnotationColumnFormatter from "./column/AnnotationColumnFormatter";
 import TumorColumnFormatter from "../mutation/column/TumorColumnFormatter";
 import SampleManager from "../sampleManager";
 import {IOncoKbData} from "../../../shared/model/OncoKB";
-import OncoKbEvidenceCache from "../OncoKbEvidenceCache";
-import PmidCache from "../PmidCache";
-import {CopyNumberCount} from "../../../shared/api/generated/CBioPortalAPIInternal";
-import MrnaExprRankCache from "../clinicalInformation/MrnaExprRankCache";
+import OncoKbEvidenceCache from "shared/cache/OncoKbEvidenceCache";
+import PmidCache from "shared/cache/PmidCache";
+import MrnaExprRankCache from "shared/cache/MrnaExprRankCache";
 import {IGisticData} from "../../../shared/model/Gistic";
 import CopyNumberCountCache from "../clinicalInformation/CopyNumberCountCache";
