+++ src/shared/cache/VariantCountCache.ts
@@ -1,6 +1,6 @@
-import LazyMobXCache from "../../../shared/lib/LazyMobXCache";
-import client from "../../../shared/api/cbioportalInternalClientInstance";
-import {VariantCount, VariantCountIdentifier} from "../../../shared/api/generated/CBioPortalAPIInternal";
+import LazyMobXCache from "shared/lib/LazyMobXCache";
+import client from "shared/api/cbioportalInternalClientInstance";
+import {VariantCount, VariantCountIdentifier} from "shared/api/generated/CBioPortalAPIInternal";
 
 function getKey<T extends { entrezGeneId:number, keyword?:string}>(obj:T):string {
     if (obj.keyword) {
